bugfix: localstorage and sessionstorage were using "Date.now() + ttl" as a expiration date but ttl was defined in seconds, so it needs to convert to milliseconds to work properly

+++ src/cache-store.js
@@ -31,7 +31,7 @@ CacheStore.prototype = {
 
   resolveTTL: function(opts) {
     var options = Utils.extend({ttl: this.ttl}, opts);
-    var ttl = parseInt(options.ttl, 10);
+    var ttl = parseFloat(options.ttl, 10);
     if (isNaN(ttl)) ttl = this.ttl;
     return ttl;
   },
